+++ CHANGELOG.md
@@ -36,6 +36,7 @@
 * Add `cast` argument to `String#assign()` method
 * Speed up the `String#underscore()` method
 * `String#decodeJSONURI()` will now only try to decode JSON if it contains certain json characters
+* Remove the `Array#createIterator()` method as it breaks CKEditor
 
 ## 0.6.6 (2019-02-25)